Features and bug fixes
Added a zoom slider to the GUI
Moved the sidebares from the borderpane to inside a stackpane on the canvas.
Fixed bugs that appered because of moving the sidebares.
Added a black style to the GUI via CSS
Created a object for the debug properties methods
Disabled the debug sidebar on startup